[feature] update machinery functionality

+++ farm_machinery/src/main/java/com/farmify/backend/dto/MachineryRequestDTO.java
@@ -32,5 +32,5 @@ public class MachineryRequestDTO {
 
     // Rotavator-specific
     private Integer bladeCount;
-    private Integer workingDepth;
+    private Double workingDepth;
 }
