+++ farm_machinery/src/main/java/com/farmify/backend/model/Rotavator.java
@@ -8,6 +8,6 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 @Data
 public class Rotavator extends Machinery {
-    private Integer workingDepth;
+    private Double workingDepth;
     private Integer bladeCount;
 }
